Reduce the need to constantly retrieve scoring HashMap
