+++ app/src/main/java/com/itachi1706/cheesecakeutilities/Modules/gpaCalculator/fragment/GpaCalcInstitutionListFragment.kt
@@ -13,11 +13,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
-import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalculatorFirebaseUtils
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.GpaCalculatorMainActivity
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.`interface`.StateSwitchListener
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaInstitution
-import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaScoring
 import com.itachi1706.cheesecakeutilities.R
 import com.itachi1706.cheesecakeutilities.RecyclerAdapters.DualLineStringRecyclerAdapter
 import com.itachi1706.cheesecakeutilities.Util.FirebaseUtils
@@ -33,7 +31,6 @@ class GpaCalcInstitutionListFragment : Fragment() {
     private val state = GpaCalculatorMainActivity.STATE_INSTITUTION
 
     private val institutions: ArrayList<GpaInstitution> = ArrayList()
-    private val scoring: HashMap<String, GpaScoring> = HashMap()
     private lateinit var adapter: DualLineStringRecyclerAdapter
 
     override fun onAttach(context: Context) {
@@ -106,7 +103,7 @@ class GpaCalcInstitutionListFragment : Fragment() {
                 }
                 LogHelper.i(TAG, "Institution Size: ${institutions.size}")
                 institutions.sortBy { it.order }
-                updateScoringTiers()
+                instituteProcessAndUpdate()
             }
         })
     }
@@ -120,29 +117,9 @@ class GpaCalcInstitutionListFragment : Fragment() {
         }
     }
 
-    private fun updateScoringTiers() {
-        val db = GpaCalculatorFirebaseUtils.getGpaDatabase().child(GpaCalculatorFirebaseUtils.FB_REC_SCORING)
-        db.keepSynced(true)
-        db.addListenerForSingleValueEvent(object: ValueEventListener{
-            override fun onCancelled(p0: DatabaseError) {
-                LogHelper.w(TAG, "updateScoringTiers:onCancelled", p0.toException())
-            }
-
-            override fun onDataChange(dataSnapshot: DataSnapshot) {
-                scoring.clear()
-                if (dataSnapshot.hasChildren()) {
-                    dataSnapshot.children.forEach{
-                        scoring[it.key!!] = it.getValue(GpaScoring::class.java)!!
-                    }
-                }
-                instituteProcessAndUpdate()
-            }
-
-        })
-    }
-
     private fun instituteProcessAndUpdate() {
         val list: ArrayList<DualLineString> = ArrayList()
+        val scoring = callback?.getScoreMap()!!
         institutions.forEach {
             val type = if (scoring.containsKey(it.type)) scoring[it.type]?.name else "Unknown"
             list.add(DualLineString("${it.name} (${it.shortName})", "Scoring Mode: $type"))
