+++ app/src/main/java/com/itachi1706/cheesecakeutilities/Modules/gpaCalculator/interface/StateSwitchListener.kt
@@ -1,6 +1,7 @@
 package com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.`interface`
 
 import com.google.firebase.database.DatabaseReference
+import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaScoring
 import com.itachi1706.cheesecakeutilities.Modules.gpaCalculator.objects.GpaSemester
 
 /**
@@ -13,5 +14,6 @@ interface StateSwitchListener {
     fun selectInstitute(instituteKey: String, instituteType: String)
     fun selectSemester(semester: GpaSemester)
     fun updateActionBar(title: String?, subtitle: String?)
+    fun getScoreMap(): HashMap<String, GpaScoring>
     fun goBack()
 }
